czoi/cli/main.py

import click
import yaml
from datetime import timedelta
from czoi.core.system import System
from czoi.core.models import Zone, Role, User, Application
from czoi.storage.sqlalchemy import Storage
from czoi.permission.engine import PermissionEngine
from czoi.simulation.engine import SimulationEngine

@click.group()
def cli():
    """CZOA Implementation Toolkit CLI."""
    pass

@cli.command()
@click.option('--config', '-c', required=True, help='YAML config file')
def init(config):
    """Initialize a new CZOA system from config."""
    with open(config) as f:
        data = yaml.safe_load(f)
    system = System()
    # Parse zones, roles, etc. from data
    # Save to database
    click.echo(f"System initialized from {config}")

@cli.command()
@click.option('--user', required=True, help='Username')
@click.option('--operation', required=True, help='Operation name')
@click.option('--zone', required=True, help='Zone name')
@click.option('--db', default='sqlite:///czoa.db', help='Database URL')
def check(user, operation, zone, db):
    """Check permission for a user."""
    storage = Storage(db)
    engine = PermissionEngine(storage)
    # Look up user, operation, zone
    # Decide and print result
    click.echo(f"Permission: ALLOWED")  # Placeholder

@cli.command()
@click.option('--db', default='sqlite:///czoa.db')
@click.option('--duration', default=60, help='Simulation duration in seconds')
@click.option('--output', default='simulation_logs.json')
def simulate(db, duration, output):
    """Run a simulation."""
    storage = Storage(db)
    system = System()  # Would load from storage
    engine = PermissionEngine(storage)
    sim = SimulationEngine(system, engine, storage)
    sim.run(timedelta(seconds=duration))
    sim.save_logs(output)
    click.echo(f"Simulation complete, logs saved to {output}")

if __name__ == '__main__':
    cli()
