requirements.txt

sqlalchemy>=1.4
click>=8.0
pyyaml>=5.4
numpy>=1.19
scikit-learn>=0.24
pytest>=7.0
